app/requirements.txt
Flask
flask-sqlalchemy
SQLAlchemy-serializer
mysqlclient